[ARM PATCH] 2033/1: IOP3xx patch submission (1/6)
Patch from Dave Jiang

This patch depends on the applied patches of 1963/1 2017/1
This is the arch/arm/boot directory

As to address RMK's previous questions in head-xscale.S, all that are workarounds
for improper things coming from redboot. Currently the distribution redboot for IOP does not pass in correct MACH ID or setup the memory map correctly after exec (according to Deepak). So we have to fix it. I'm working on pushing the IOP redboot developer to fix those things at a later date.

BKrev: 413b65a1iC5rrxoJBj6OYqDJ30NZ4Q