Merge bk://bk.arm.linux.org.uk/linux-2.6-mmc into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/09/06 19:36:50+01:00 rmk
[MMC] MMC_RSP_xxx combined response types.

Create a set of MMC_RSP_xxx definitions which correspond to the
response type we expect at a high level.  These are created from
a set of flags; a MMC host driver can either interpret the flags
as the high level response type, or the low level individual
flags.

2004/08/30 17:43:10+01:00 rmk
[MMC] PXAMCI: Bracket power management calls with CONFIG_PM.

2004/08/30 17:33:23+01:00 rmk
[MMC] MMCI: Ensure that we read all data from FIFO.

When we are nearing the end of a transfer from the card, we may not
have sufficient data in the FIFO to trigger a "half FIFO full" IRQ.
Therefore, when we have less than MCI_FIFOSIZE bytes left, switch to
"FIFO contains data" IRQ mode instead.

2004/08/30 17:17:25+01:00 rmk
[MMC] MMCI: Add SG support to PIO data transfers

This allows us to transfer a whole request for multiple pages in one
transaction with the MMC card, resulting in higher overall throughput.

2004/08/30 17:03:48+01:00 rmk
[MMC] MMCI: use bio_kmap_irq() rather than req->buffer

This is in preparation for SG data IO.

2004/08/30 16:23:08+01:00 rmk
[MMC] MMCI: Maintain offset rather than buffer pointer for PIO

2004/08/30 16:13:28+01:00 rmk
[MMC] MMCI: split the PIO data read and write paths.

2004/08/30 15:38:01+01:00 rmk
[MMC] MMCI: Manipulate IRQ masks according to the data FSM state.

Prevent IRQ race conditions between the two handlers and within the
primecell itself.  Unfortunately, the primecell does not always mask
the FIFO empty interrupt we reach the end of a transfer to the card.
Also, we may receive the "data end" interrupt prior to finishing a
transfer from the card, so mask this off until we've read the last
bytes from the FIFO.

2004/08/30 12:46:45+01:00 rmk
[MMC] MMCI: data FSM handling updates

- Use mmci_stop_host() to ensure that the data FSM is shut down and
  any driver held data is properly released.
- FIFO size is 64 bytes not 16.

2004/08/30 11:12:18+01:00 rmk
[MMC] MMCI: Remove hardcoded MCI_IRQMASK definition

2004/08/30 10:58:22+01:00 rmk
[MMC] Give the MMC host the full-sized request.

Pass the total number of MMC data blocks in the request to the
host driver.  Since existing MMC host drivers do not support
scatter-gather at present, the block queue defaults have been
chosen to prevent scatter-gather block requests for MMC.  A
host can alter these parameters when it wishes to take advantage
of SG capabilities.

2004/08/30 10:49:37+01:00 rmk
[MMC] Add host specific block queue parameters

Allow MMC host drivers to specify their data phase capabilities
to the block queue.  This will eventually allow us to perform
SG data phase operations, thereby processing a single request
in one go.

2004/08/30 10:30:55+01:00 rmk
[MMC] Remove unused host->priv

2004/08/30 10:20:35+01:00 rmk
[MMC] Use local card pointer rather than md->queue.card

BKrev: 413e4732pcmwbOIn2UDSjffpHrwwqA