[PATCH] uml: Avoid forcing use of the no-op scheduler
Avoid forcing use of the no-op scheduler for UBD; this may uncover some
bugs in the UBD driver, and in fact uml-ubd-no-empty-queue.patch is needed
to make this sure.  But as of now, no other bugs have been discovered, so
this should be safe.

Signed-off-by: Paolo 'Blaisorblade' Giarrusso <blaisorblade_spam@yahoo.it>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 413f1e53mzt5SJs03XWn9IpNrjkWJw