[PATCH] ppc/ppc64: fix offb
The recent fbdev changes broke something quite fundamental on ppc, which
is the way offb works. It's a "fallback" driver that is to be used when
no other driver picked up the video, or when "forced" via the video=ofonly
command line option.

The recent changes completely broke that (which is a pretty important
behaviour on ppc since we still have some models regulary with video
cards that don't work properly with the kernel drivers, like some nVidias
or some older stuffs).

This patch moves offb to the end of the Makefile, so at least the behaviour
of taking over as a "fallback" is restored (the current kernel will have
offb take control of the framebuffer before any fbdev has a chance to do
it, which breaks everything).

Apparently, the entire support for "video=ofonly" was removed though,
that NEEDS to be restored in some way, though i'm not yet sure what is
the best path to that yet, I have to look more deeply at the new code.

BKrev: 4141c682rvTLrAYJCxvrb6Nccl_AGg