[NETFILTER]: Fix conntrack seq_file handling.
Am travelling, but this passed simple tests here.  If this isn't going
in, the current seqfile stuff should be ripped out; it's a mess.

/proc/net/ip_conntrack was changed over to seq_file.  However,
seq_file isn't a great fit (a linked list which is changing is not a
good candidate for seq file), and the conversion was done badly.

1) Don't do allocation: simply hand the pointer head of the correct chain.
2) Actually output the original tuple.
3) Lock only when actually traversing hash chain.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

BKrev: 4144e32fVuALpXBYri9y6VOz5sHjCg