[PATCH] ppc64: enable POWER5 low power mode in idle loop
Now that we understand (and have fixed) the problem with using low power mode
in the idle loop, lets enable it.  It should save a fair amount of power.

(The problem was that our exceptions were inheriting the low power mode and so
were executing at a fraction of the normal cpu issue rate.  We fixed it by
always bumping our priority to medium at the start of every exception).

Signed-off-by: Anton Blanchard <anton@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4147037alSk3B52_Zub6WQXV72vpvw