[PATCH] __copy_to_user() check in cdrom_read_cdda_old()
akpm: really, reads are supposed to return the number-of-bytes-read on faults,
or -EFAULT of no bytes were read.  This patch returns either zero or -EFAULT,
ignoring any successfully transferred data.  But the user interface (whcih is
an ioctl()) was never set up to do that.

Signed-off-by: Jesper Juhl <juhl-lkml@dif.dk>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41470466rd7VRhd4GBV_eTNB854cdA