[PATCH] i386: elf_read_implies_exec() fixup
The final ia64 related cleanup to elf_read_implies_exec() seems to have
broken it.  The effect is that the READ_IMPLIES_EXEC flag is never set
for !pt_gnu_stack binaries!

That's a bit more secure than we need to be, and might break some legacy
app that doesn't expect it.

BKrev: 41471440blw23uXoDRUhA8jYMd16fw