Merge http://linux-mh.bkbits.net/bluetooth-2.6 into nuts.davemloft.net:/disk1/BK/net-2.6
2004/09/08 11:51:10+02:00 marcel
[Bluetooth] Check checksums for BNEP

The checking of the checksums is needed, because with certain Bluetooth
hardware there are failures which aren't caught otherwise.

Signed-off-by: David Woodhouse <dwmw2@infradead.org>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/09/04 12:52:38+02:00 marcel
[Bluetooth] Don't send L2CAP reject command for bad responses

If a L2CAP response is bad or not understood there is no need to send
a reject command. The right way is to ignore the response and let the
L2CAP timers trigger the further actions.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/09/04 12:42:34+02:00 marcel
[Bluetooth] Don't use ISOC transfers for Broadcom dongle

The Broadcom dongles with HID proxy support don't work with SCO over
HCI and so there is no need to start the ISOC transfers.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

BKrev: 4149f793Bf5kr1PjDk5MMQLfZoMUDQ