[PATCH] ppc32: Fix arch/ppc/boot/common/ns16550.c
When <linux/timex.h> started including <asm/io.h> this exposed one of the
fragilities of the code in arch/ppc/boot/, namely that it is tied to the
kernel headers for some information, yet not really the kernel.  The
following starts us in the direction of being less tied to the kernel by
providing our own serial_state definition (all we care about is the ability
to grab information from SERIAL_PORT_DFNS).

Signed-off-by: Tom Rini <trini@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 414b333fBxxqQmmj1gDTDURYTnoeIQ