[PATCH] ppc64: replace mmu_context_queue with idr allocator
Replace the mmu_context_queue structure with the idr allocator.  The
mmu_context_queue allocation was quite large (~200kB) so on most machines
we will have a reduction in usage.

We might put a single entry cache on the front of this so we are more
likely to reuse ppc64 MMU hashtable entries that are in the caches.

Signed-off-by: Anton Blanchard <anton@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 414b3372ND2qfXWmdIo4yY8QSIbL-g