Merge davem@nuts.davemloft.net:/disk1/BK/net-2.6 into kernel.bkbits.net:/home/davem/net-2.6
2004/09/17 23:20:35-07:00 herbert
[IPV4]: Size fib_info_devhash[] correctly.

There is a thinko in the allocation for the devindex hash.  We're
only giving it 8 elements when it should be 1<<8 elements.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2004/09/17 23:17:16-07:00 davem
[IPV4]: Fix BUG triggered in fib_sync_down()

If you modload then unload a network device,
without creating any ipv4 routes, it is possible
for the laddr hash to be not allocated yet
in fib_semantics.

Signed-off-by: David S. Miller <davem@davemloft.net>

2004/09/17 15:17:53-07:00 davem
[XFRM] make xfrm_lookup() fully af-independent.

Simplified from 2 patches by Hideaki YOSHIFUJI
<yoshfuji@linux-ipv6.org>

Signed-off-by: David S. Miller <davem@davemloft.net>

BKrev: 414c3a535dwn2GbYcAdN1iZUcH-YAg