Merge bk://bk.arm.linux.org.uk/linux-2.6-mmc into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/09/18 17:47:23+01:00 rmk
[MMC] Add v2.x and v3.x CID parsing.

2004/09/18 16:32:07+01:00 rmk
[MMC] Clean up MMC card CID/CSD decoding, stage 1.

Now that the raw CSD and CID information are available in mmc_card,
there's little point not passing the mmc_card structure into the
decode functions themselves.

This will be necessary, since there's several different formats of
CID decode, and the format appears to depend on data in the CSD.

If we are unable to decode the CID or CSD, we can mark the card bad
and refuse to bind it to any drivers.  Users can then retrieve the
raw undecoded CID and CSD from sysfs for debugging purposes.

2004/09/18 16:06:51+01:00 rmk
[MMC] Export raw MMC card CID and CSD registers via device model.

2004/09/18 15:38:44+01:00 rmk
[MMC] Save MMC card raw CSD structure.

2004/09/18 10:50:57+01:00 rmk
[MMC] Use raw CID rather than decoded CID

Save the raw CID information, and use this to match cards
with their corresponding mmc_card structure.  Different
protocol revisions have different CID formats.

BKrev: 414cfc87I_BLkFhpaCWbT_LuLlW3-A