Merge bk://bk.arm.linux.org.uk/linux-2.6-mmc into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/09/20 21:11:41+01:00 rmk
[MMC] Ensure semaphores are initialised before use.  Gah.

2004/09/20 20:44:59+01:00 rmk
[MMC] PXAMCI: enable use of platform specific data.

This allows platforms to hook in their card detection, power control
and power availability information.

2004/09/20 10:21:45+01:00 rmk
[MMC] Ensure PXA MMC interrupts are disabled on chip.

2004/09/20 10:07:54+01:00 rmk
[MMC] Fix mmc_block suspend/resume handling (again).

The previous change is not the whole story - the mmc queue
thread may be handling a request.  We must wait for outstanding
requests to complete before allowing the suspend to proceed,
otherwise we may suffer loss of data.

BKrev: 41503d0cdeKUSyJnSeBirMg4bFTG2w