[PATCH] Warn that ipchains and ipfwadm are going away
At the recent netfilter workshop in Erlangen, we was decided to remove
the backwards compatibility code for ipchains and ipfwadm.  This will
allow significant cleanup of interfaces, since we had to have a
mid-level interface for the backwards compatibility layer to use.

Start off with a warning for 2.6.9, so any remaining users have a chance
to migrate.  Their firewall scripts might not check return values, and
they might get a nasty surprise when this goes away.

BKrev: 41503d5fEbnONPiJ-H0MxnUpVn74Ag