[PATCH] Online CPU with maxcpus option panics
I booted IA64 kernel with "maxcpus=" set to less than the total available
processors in the system.  Then I tried to online the CPU which was offline
due to "maxcpus=" restirction.  i.e when I tried to online the CPU which
never was online at boot crashed the system.

Signed-off-by: Anil Keshavamurthy <anil.s.keshavamurthy@intel.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4152250fCzOveH3pk7ndvqYQJfwXtQ