[PATCH] ppc32: Fix potentially uninitialized var in chrp_setup.c
In chrp_setup.c, chrp_int_ack could be left uninitialized and passed
"as-is" to i8259_init() if the OF node for the 8259 wasn't found.  I
don't know if that should happen, but the i8259 code can deal with
int_ack beeing 0, so let's be safe, initialize it, and remove a warning
at the same time.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 415374cc4_-sz_3PyiK1ywg_8_m7NA