Merge bk://linux-ntfs.bkbits.net/ntfs-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/09/24 08:59:31-07:00 jbarnes
[PATCH] disable sched domains debug code

This disables the sched domains debug code, which has a tendency to
print out a *lot* of information.  It's helpful for debugging, but a bit
overboard for general use.  Also, if you have a high NR_CPUS value, it
tends to scare people and overflow the dmesg buffer.

Signed-off-by: Nick Piggin <nickpiggin@yahoo.com.au>
Signed-off-by: Jesse Barnes <jbarnes@sgi.com>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41544bf9U8ly7-0sS0oAx_jTmlPBbw