Merge bk://are.twiddle.net/axp-2.6/ into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/09/25 18:41:38-07:00 benh
[PATCH] ppc64: Fix 32 bits conversion of SI_TIMER signals

The current 32 bits translation of the SI_TIMER is wrong on ppc64, causing
the tst-timer4 testcase of glibc to fail in 32 bits. This patch fixes it.

Signed-off-by: Olaf Hering <olh@suse.de>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

2004/09/25 18:35:17-07:00 willy
[PATCH] fix posix_locks_deadlock()

The "blocked_list" used for POSIX deadlock detection should only contain
POSIX locks, or posix_locks_deadlock() can get confused.

As I've noted before (but never posted a patch), we should probably stop
checking for deadlock.  POSIX says it's optional for an implementation
to check for deadlock, and since the change to match on tgid, we can
report a deadlock when none exists.

[ Further explanation for Linus: POSIX locks and BSD locks don't block
  each other, so checking that the blocker is a POSIX lock also implies
  that the waiter is a POSIX lock. ]

2004/09/25 15:51:42-07:00 torvalds
Make smbfs with UNIX extensions get file disk usage count right.

We need to convert from bytes to blocks, and we also need to
handle the fact that old smbd's will round up the disk usage
bytecount to even multiples of 1MB, and gracefully fall back
on an estimate based on the file size instead in that case.

Acked by Samuel Thibault who noted the problem in the first
place, and Jeremy Allison wrt smbd behavior.

Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41562b6fjjxeSgQNTihLjfiosNh0Rg