[PATCH] bttv: fix DMA setup bug in latest update
This fixes the bug that made the card DMA lots of data to random memory
locations, causing lockups and instant reboots.

The problem was that the yoffset variable got modified inside the loop,
but the logic in the switch statement was meant to work on the initial
value of the yoffset variable.

(Bug fix extracted from
http://marc.theaimsgroup.com/?l=linux-kernel&m=109532814823565&w=2)

Signed-off-by: Peter Osterlund <petero2@telia.com>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4156f331EPczBHB8fNXefNzfgMfXag