[PATCH] Add wait_event_timeout()
There appears to be one case missing from the wait_event() family - the
uninterruptible timeout wait.  The following patch adds this.

This wait is particularly useful when (eg) you wish to pass work off to an
interrupt handler to perform, but also want to know if the hardware has
decided to go gaga under you.  You don't want to sit around waiting for
something that'll never happen - you want to go and wack the gremlin which
caused the failure and retry.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41582257b0ChL5S5rs1Hb9vZ40xL2A